Moved ProcBlockNode and ModelNode to their own modules
